+++ packages/xchain-cosmos/CHANGELOG.md
@@ -42,10 +42,10 @@
 
 ### Update
 
-- Update to latest `@xchainjs/*` packages and other dependencies
+- Update to latest `@bithighlander/*` packages and other dependencies
 
 # v.0.4.0 (2020-11-20)
 
 ### Breaking change
 
-- Update @xchainjs/xchain-crypto package to 0.2.0, deprecating old keystores
+- Update @bithighlander/xchain-crypto package to 0.2.0, deprecating old keystores