+++ packages/xchain-thorchain/CHANGELOG.md
@@ -78,16 +78,16 @@
 
 ### Update
 
-- Use latest `@xchainjs/cosmos@0.4.2`
+- Use latest `@bithighlander/cosmos@0.4.2`
 
 # v.0.4.1 (2020-11-23)
 
 ### Update
 
-- Update to latest `@xchainjs/*` packages and other dependencies
+- Update to latest `@bithighlander/*` packages and other dependencies
 
 # v.0.4.0 (2020-11-20)
 
 ### Breaking change
 
-- Update @xchainjs/xchain-crypto package to 0.2.0, deprecating old keystores
+- Update @bithighlander/xchain-crypto package to 0.2.0, deprecating old keystores